requirements.txt
BeautifulSoup4
google-generativeai
numpy
openai
pandas
PyMuPDF
python-docx
python-dotenv
requests
scikit-learn
sentence-transformers
snowflake-connector-python
streamlit
tiktoken
